Split out StripePayment component to its own file
